# src/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This ios Intellidrone API designed by [Singleton](http://singleton.com.ar).You\
    \  can find out more about Swagger at [Singleton Web Page](http://singleton.com.ar)\
    \ or on [Intellidrone Web Page](http://robotagro.com)."
  version: "1.0.0"
  title: "Intellidrone API"
  termsOfService: "http://web.robotagro.com/api//terms"
  contact:
    name: "Equipo de Soporte"
    url: "http://singleton.com.ar/soporte"
    email: "soporte@singleton.com.ar"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
- "https"
paths:
...
  /Users:
    get:
      tags:
      - "Users"
      summary: "Get all Users."
      description: "Get all Users."
      operationId: "getUsers"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "skip"
        in: "query"
        description: "number of item to skip"
        required: true
        type: "integer"
        format: "int32"
      - name: "limit"
        in: "query"
        description: "max records to return"
        required: true
        type: "integer"
        format: "int32"
      - name: "orderBy"
        in: "query"
        description: "order by property."
        required: false
        type: "string"
      - name: "filter"
        in: "query"
        description: "filter data."
        required: false
        type: "string"
      - name: "deleted"
        in: "query"
        description: "Get all, daleted, not deleted data. Default not deleted."
        required: false
        type: "string"
        default: "NOT-DELETED"
        enum:
        - "NOT-DELETED"
        - "DELETED"
        - "ALL"
      - name: "userId"
        in: "query"
        description: "id of user. Only for admin, editor or viewer users."
        required: false
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Users"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Users"
    post:
      tags:
      - "Users"
      summary: "Add one User."
      description: "Add one User."
      operationId: "addUser"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "User"
        required: true
        schema:
          $ref: "#/definitions/Users"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Users"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Users"
    put:
      tags:
      - "Users"
      summary: "Edit one User."
      description: "Edit one User."
      operationId: "editUser"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "body"
        name: "User"
        required: true
        schema:
          $ref: "#/definitions/Users"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Users"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Users"
  /Users/{id}:
    get:
      tags:
      - "Users"
      summary: "Get one User."
      description: "Get one User."
      operationId: "getUserById"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id"
        in: "path"
        description: "id to delete"
        required: true
        type: "integer"
        format: "int64"
      - name: "deleted"
        in: "query"
        description: "Get all, daleted, not deleted data. Default not deleted."
        required: false
        type: "string"
        default: "NOT-DELETED"
        enum:
        - "NOT-DELETED"
        - "DELETED"
        - "ALL"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Users"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Users"
    delete:
      tags:
      - "Users"
      summary: "Delete one Users."
      description: "Delete one User."
      operationId: "deleteUsers"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - name: "id"
        in: "path"
        description: "id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Request OK."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/DeletedResponse"
        204:
          description: "Response is empty."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        400:
          description: "Some parameters are missing or badly entered."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized"
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "string"
              format: "date-time"
              description: "The UTC date/time at which the current rate limit window\
                \  resets."
          schema:
            type: "string"
        404:
          description: "Entity not found."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
        405:
          description: "Illegal input for operation."
          headers:
            X-Rate-Limit-Limit:
              type: "integer"
              description: "The number of allowed requests in the current period"
            X-Rate-Limit-Remaining:
              type: "integer"
              description: "The number of remaining requests in the current period"
            X-Rate-Limit-Reset:
              type: "integer"
              description: "The number of seconds left in the current period"
          schema:
            $ref: "#/definitions/Error"
      security:
      - appToken: []
      - userToken: []
      x-swagger-router-controller: "Users"
...
securityDefinitions:
  appToken:
    type: "apiKey"
    name: "x_app_id"
    in: "header"
  userToken:
    type: "apiKey"
    name: "x_user_key"
    in: "header"
definitions:
  Lotes:
    type: "object"
    properties:
      _id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      latitude:
        type: "number"
        format: "float"
      longitude:
        type: "number"
        format: "float"
    example:
      latitude: 6.0274563
      name: "Campo Baires"
      id: 0
      longitude: 1.4658129
...
  Users:
    type: "object"
    properties:
      _id:
        type: "integer"
        format: "int64"
      user:
        type: "string"
      password:
        type: "string"
        format: "password"
      lotes:
        type: "array"
        items:
          $ref: "#/definitions/Lotes"
      type:
        type: "string"
        enum:
        - "COMMON"
        - "SUPERVISOR"
        - "ADMIN"
    example:
      _id: 0
      type: "COMMON"
      user: "user"
      password: "Password"
...
  Error:
    type: "object"
    required:
    - "message"
    properties:
      code:
        type: "string"
      message:
        type: "string"
      data:
        type: "string"
    example:
      code: "code"
      data: "data"
      message: "message"
  DeletedResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
    example:
      id: 0